split up trial runner test
